Updated geoPlotGini
The function now allows the user to specify an optional vector of crime
numbers per source and returns Gini plots and data for both proportion
of suspects and proportion of crimes. Note that crimeNumbers needs to
be the same length as the sources, and in the same order as the
original list of sources, although it is returned sorted.